requirements.txt
requests
beautifulsoup4
discord.py
datetime